pyproject.toml
[project]
name = "BSc Thesis"
description = "BSc Thesis"
version = "1.0.0"
authors = [
    { name = "Jonas Hoffmann", email = "s204071@student.dtu.dk" },
    { name = "Joachim S. Andersson", email = "s204096@student.dtu.dk" }
]
readme = "README.md"
requires-python = ">=3.9"

[tool.poetry]
name = "bsc_metadata_archeology"
version = "1.0.0"
description = ""
authors = ["Jonas Hoffmannn <s204071@student.dtu.dk>", "Joachim S. Andersson <s204096@student.dtu.dk>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
sphinx = "^6.1.3"
coverage = "^7.1.0"
python-dotenv = "^0.21.1"
torch = "^1.13.1"
torchvision = "^0.14.1"
torchmetrics = "^0.11.1"
pytorch-lightning = "^1.9.1"
numpy = "^1.24.2"
matplotlib = "^3.6.3"
dtuhpc = "^2.1.0"
tqdm = "^4.64.1"
hydra-core = "^1.3.1"
wandb = "^0.13.10"
pre-commit = "^3.0.4"
pyarrow = "^11.0.0"
scikit-learn = "^1.2.1"
pandas = "^1.5.3"
seaborn = "^0.12.2"
requests = "^2.28.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
black = "^23.1.0"
pre-commit = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 88
